+++ packages/@storefront/flux-capacitor/src/core/adapters/past-purchases.ts
@@ -1,6 +1,7 @@
 import { Biasing, Sort } from 'groupby-api';
 import Selectors from '../selectors';
 import Store from '../store';
+import { PAST_PURCHASE_SORTS } from '../utils';
 
 namespace PastPurchasesAdapter {
 
@@ -33,6 +34,14 @@ namespace PastPurchasesAdapter {
     return [...skus].sort(({ [field]: lhs }, { [field]: rhs }) => rhs - lhs);
   };
 
+  export const sortSkusByField = (skus: Store.PastPurchases.PastPurchaseProduct[], field?: string) => {
+    switch (field) {
+      case PAST_PURCHASE_SORTS.MOST_PURCHASED: return PastPurchasesAdapter.sortSkus(skus, 'quantity');
+      case PAST_PURCHASE_SORTS.MOST_RECENT: return PastPurchasesAdapter.sortSkus(skus, 'lastPurchased');
+      default: return skus;
+    }
+  };
+
   export const biasSkus = (state: Store.State) => {
     const ids: string[] = Selectors.pastPurchases(state).map(({ sku }) => sku);
 
@@ -42,7 +51,7 @@ namespace PastPurchasesAdapter {
       },
       sort: <Sort[]>[{ type: 'ByIds', ids }],
     };
-  }
+  };
 }
 
 export default PastPurchasesAdapter;