+++ packages/@storefront/flux-capacitor/src/core/adapters/request.ts
@@ -1,4 +1,5 @@
 import { SelectedRefinement, Sort } from 'groupby-api';
+import { PAST_PURCHASE_SORTS } from '../utils';
 import Store from '../store';
 import { MAX_RECORDS } from './search';
 
@@ -11,6 +12,12 @@ namespace RequestAdapter {
 
   export const extractSort = ({ field, descending }: Store.Sort): Sort =>
     ({ field, order: descending ? 'Descending' : undefined });
+
+  // tslint:disable-next-line max-line-length
+  export const extractPastPurchaseSort = ({ field, descending }: Store.Sort, skus: Store.PastPurchases.PastPurchaseProduct[]): Sort =>
+    Object.keys(PAST_PURCHASE_SORTS).map((k) => PAST_PURCHASE_SORTS[k]).includes(field)
+      ? { type: 'ByIds', ids: skus.map(({ sku }) => sku) }
+      : RequestAdapter.extractSort({ field, descending });
 
   export const extractRefinement = (field: string, refinement: Store.Refinement): SelectedRefinement =>
     (<Store.ValueRefinement>refinement).value