+++ packages/@storefront/flux-capacitor/src/core/configuration.ts
@@ -113,6 +113,14 @@ namespace Configuration {
     sayt: Sayt;
   }
 
+  export interface Sort {
+    /**
+     * field path to sort on
+     */
+    field: string;
+    descending?: boolean;
+  }
+
   export interface Search {
     /**
      * product fields to request
@@ -126,13 +134,7 @@ namespace Configuration {
     /**
      * sorting of products or sort options and default
      */
-    sort?: Configuration.ValueOptions<{
-      /**
-       * field path to sort on
-       */
-      field: string;
-      descending?: boolean;
-    }>;
+    sort?: Configuration.ValueOptions<Configuration.Sort>;
     /**
      * redirect to the details page of product if there is only 1 product result for a search
      */
@@ -336,6 +338,11 @@ namespace Configuration {
        * A security token or a function to retrieve a token
        */
       securedPayload: Recommendations.SecuredPayload | (() => Recommendations.SecuredPayload);
+
+      /**
+       * sorting of products or sort options and default
+       */
+      sort?: Configuration.ValueOptions<Configuration.Sort>;
 
       /**
        * Enable past purchases or not