+++ packages/@storefront/flux-capacitor/src/core/index.ts
@@ -9,11 +9,12 @@ import { DEFAULT_AREA } from './reducers/data/area';
 import { DEFAULT_COLLECTION } from './reducers/data/collections';
 import Selectors from './selectors';
 import Store, { ReduxStore } from './store';
-import { Routes, StoreSections } from './utils';
+import { PAST_PURCHASE_SORTS, Routes, StoreSections } from './utils';
 
 export {
   DEFAULT_AREA,
   DEFAULT_COLLECTION,
+  PAST_PURCHASE_SORTS,
   ActionCreators,
   Actions,
   Adapters,