+++ packages/@storefront/flux-capacitor/src/core/requests/utils.ts
@@ -1,4 +1,4 @@
-import { Request } from 'groupby-api';
+import { FieldSort, Request } from 'groupby-api';
 import { QueryTimeAutocompleteConfig } from 'sayt';
 import Autocomplete from '../adapters/autocomplete';
 import Configuration from '../adapters/configuration';
@@ -59,7 +59,10 @@ namespace RequestHelpers {
       query: Selectors.pastPurchaseQuery(state),
       refinements: Selectors.pastPurchaseSelectedRefinements(state),
       skip: Selectors.pastPurchasePageSize(state) * (Selectors.pastPurchasePage(state) - 1),
-      sort: RequestAdapter.extractSort(Selectors.pastPurchaseSortSelected(state)),
+      sort: RequestAdapter.extractPastPurchaseSort(
+        Selectors.pastPurchaseSortSelected(state),
+        Selectors.pastPurchases(state)
+      ),
     };
 
     return <Request>{ ...request, ...overrideRequest };