+++ packages/@storefront/core/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] [patch]
+### Fixed
+- SF-1237: Fix issue where past purchase sorts are not correctly applied/transmitted.
+
 ## [1.55.5] - 2019-01-04
 ### Changed
 - Update `@storefront/flux-capacitor` to 1.67.3.